+++ CLIENT8106/src/main/java/es/wuolahpop/data/TransactionFromClient.java
@@ -3,14 +3,14 @@ package es.wuolahpop.data;
 public class TransactionFromClient {
 	private String comprador;
 	private String vendedor;
-	private Long coste;
+	private float coste;
 	private String fechaTarjeta;
 	private String fechaTransaccion;
 	int cv2;
 	Long tarjeta;
 	private String codigoPedido;
 	
-	public TransactionFromClient(String comprador, String vendedor, Long coste, String fechaTarjeta,
+	public TransactionFromClient(String comprador, String vendedor, float coste, String fechaTarjeta,
 			String fechaTransaccion, int cv2, Long tarjeta, String codigoPedido) {
 		super();
 		this.comprador = comprador;
@@ -35,7 +35,7 @@ public class TransactionFromClient {
 	public void setVendedor(String vendedor) {
 		this.vendedor = vendedor;
 	}
-	public Long getCoste() {
+	public float getCoste() {
 		return coste;
 	}
 	public void setCoste(Long coste) {
